fix: refactored rpc functionality
